パラメータのエフェクト側からの読み取りに大体対応
git-svn-id: http://picopicose.com/svn/C700@66 38616c85-8a89-495d-9829-d15a903298fa

+++ WaveView.h
@@ -23,7 +23,7 @@ public:
 	virtual void onDragMove (CDragContainer* drag, const CPoint& where);
 	
 	void draw(CDrawContext *pContext);
-	void setWave(float *wavedata, long frames);
+	void setWave(short *wavedata, long frames);
 	void setBackColor(CColor color);
 	void setLineColor(CColor color);
 	bool isActive(void) {return isWaveLoaded;}
